infared sensors returns varying directions

+++ main/infrared.hpp
@@ -23,7 +23,7 @@ class InfraredSensorsTriad {
       const InfraredSensor &sensorInfraredC,
       const InfraredSensor &sensorInfraredR);
   InfraredState* read();
-  int getDirection();
+  float getDirection();
  private:
   // How??? https://stackoverflow.com/questions/47897253/error-no-matching-function-to-call-for-with-class-as-attribute-c
   InfraredSensor sensorInfraredL;
